requirements.txt
pandas
geopandas
numpy
matplotlib
seaborn
# gdal conda-forgey
rasterio
shapely
cdsapi
elevation
overpy
shapely
geopy
openmeteo_requests
requests_cache
retry_requests